Refactor Unspent domain:  * Reorder mehtods & Rename method UnLock to Unlock  * Reorder repo methods & Rename method GetUnspentForKey to GetUnspentWithKey  * Reorganize repo error variables  * Move BalanceInfo type  to app level  * Apply changes to other aprts of the codebase accordingly
