+++ internal/core/domain/unspent_model.go
@@ -2,15 +2,17 @@ package domain
 
 import (
 	"github.com/google/uuid"
-	"github.com/tdex-network/tdex-daemon/pkg/explorer"
-	"github.com/tdex-network/tdex-daemon/pkg/explorer/esplora"
 )
 
+// UnspentKey represent the ID of an Unspent, composed by its txid and vout.
 type UnspentKey struct {
 	TxID string
 	VOut uint32
 }
 
+// Unspent is the data structure representing an Elements based UTXO with some
+// other information like whether it is spent/unspent, confirmed/unconfirmed or
+// locked/unlocked.
 type Unspent struct {
 	TxID            string
 	VOut            uint32
@@ -29,69 +31,4 @@ type Unspent struct {
 	Locked          bool
 	LockedBy        *uuid.UUID
 	Confirmed       bool
-}
-
-type BalanceInfo struct {
-	TotalBalance       uint64
-	ConfirmedBalance   uint64
-	UnconfirmedBalance uint64
-}
-
-func (u *Unspent) Lock(tradeID *uuid.UUID) {
-	u.Locked = true
-	u.LockedBy = tradeID
-}
-
-func (u *Unspent) UnLock() {
-	u.Locked = false
-	u.LockedBy = nil
-}
-
-func (u *Unspent) IsLocked() bool {
-	return u.Locked
-}
-
-func (u *Unspent) Spend() {
-	u.Spent = true
-}
-
-func (u *Unspent) Confirm() {
-	u.Confirmed = true
-}
-
-func (u *Unspent) IsSpent() bool {
-	return u.Spent
-}
-
-func (u *Unspent) IsConfirmed() bool {
-	return u.Confirmed
-}
-
-func (u *Unspent) Key() UnspentKey {
-	return UnspentKey{
-		TxID: u.TxID,
-		VOut: u.VOut,
-	}
-}
-
-func (u *Unspent) IsKeyEqual(key UnspentKey) bool {
-	return u.TxID == key.TxID && u.VOut == key.VOut
-}
-
-func (u *Unspent) ToUtxo() explorer.Utxo {
-	return esplora.NewWitnessUtxo(
-		u.TxID,
-		u.VOut,
-		u.Value,
-		u.AssetHash,
-		u.ValueCommitment,
-		u.AssetCommitment,
-		u.ValueBlinder,
-		u.AssetBlinder,
-		u.ScriptPubKey,
-		u.Nonce,
-		u.RangeProof,
-		u.SurjectionProof,
-		u.Confirmed,
-	)
 }
