+++ internal/infrastructure/storage/db/badger/unspent_repository_impl.go
@@ -45,15 +45,6 @@ func (u unspentRepositoryImpl) GetAllUnspents(
 	return u.getAllUnspents(ctx)
 }
 
-func (u unspentRepositoryImpl) GetBalance(
-	ctx context.Context,
-	addresses []string,
-	assetHash string,
-) (uint64, error) {
-	unlockedOnly := false
-	return u.getBalance(ctx, addresses, assetHash, unlockedOnly)
-}
-
 func (u unspentRepositoryImpl) GetAvailableUnspents(
 	ctx context.Context,
 ) ([]domain.Unspent, error) {
@@ -132,6 +123,22 @@ func (u unspentRepositoryImpl) GetAvailableUnspentsForAddresses(
 	return unspents, nil
 }
 
+func (u unspentRepositoryImpl) GetUnspentWithKey(
+	ctx context.Context,
+	unspentKey domain.UnspentKey,
+) (*domain.Unspent, error) {
+	return u.getUnspent(ctx, unspentKey)
+}
+
+func (u unspentRepositoryImpl) GetBalance(
+	ctx context.Context,
+	addresses []string,
+	assetHash string,
+) (uint64, error) {
+	unlockedOnly := false
+	return u.getBalance(ctx, addresses, assetHash, unlockedOnly)
+}
+
 func (u unspentRepositoryImpl) GetUnlockedBalance(
 	ctx context.Context,
 	addresses []string,
@@ -167,13 +174,6 @@ func (u unspentRepositoryImpl) UnlockUnspents(
 	ctx context.Context,
 	unspentKeys []domain.UnspentKey) error {
 	return u.unlockUnspents(ctx, unspentKeys)
-}
-
-func (u unspentRepositoryImpl) GetUnspentForKey(
-	ctx context.Context,
-	unspentKey domain.UnspentKey,
-) (*domain.Unspent, error) {
-	return u.getUnspent(ctx, unspentKey)
 }
 
 func (u unspentRepositoryImpl) addUnspents(
@@ -284,7 +284,7 @@ func (u unspentRepositoryImpl) spendUnspent(
 	}
 
 	unspent.Spend()
-	unspent.UnLock() // prevent conflict, locks not stored under unspent prefix
+	unspent.Unlock() // prevent conflict, locks not stored under unspent prefix
 
 	return u.updateUnspent(ctx, key, *unspent)
 }
@@ -315,7 +315,7 @@ func (u unspentRepositoryImpl) confirmUnspent(
 	}
 
 	unspent.Confirm()
-	unspent.UnLock() // prevent conflict, locks not stored under unspent prefix
+	unspent.Unlock() // prevent conflict, locks not stored under unspent prefix
 
 	return u.updateUnspent(ctx, key, *unspent)
 }
