+++ internal/infrastructure/storage/db/badger/unspent_repository_impl_test.go
@@ -150,6 +150,7 @@ func TestGetAllUnspentsForAddresses(t *testing.T) {
 	assert.Equal(t, 4, len(unspents))
 
 }
+
 func TestGetUnspentsForAddresses(t *testing.T) {
 	before()
 	defer after()
@@ -178,11 +179,11 @@ func TestGetAvailableUnspentsForAddresses(t *testing.T) {
 	assert.Equal(t, 2, len(unspents))
 }
 
-func TestGetUnspentForKey(t *testing.T) {
+func TestGetUnspentWithKey(t *testing.T) {
 	before()
 	defer after()
 
-	unspent, err := unspentRepository.GetUnspentForKey(ctx, domain.UnspentKey{
+	unspent, err := unspentRepository.GetUnspentWithKey(ctx, domain.UnspentKey{
 		TxID: "1",
 		VOut: 1,
 	})
@@ -201,7 +202,7 @@ func TestSpendUnspents(t *testing.T) {
 		TxID: "1",
 		VOut: 1,
 	}
-	unspent, err := unspentRepository.GetUnspentForKey(ctx, unspentKey)
+	unspent, err := unspentRepository.GetUnspentWithKey(ctx, unspentKey)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -215,7 +216,7 @@ func TestSpendUnspents(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	unspent, err = unspentRepository.GetUnspentForKey(ctx, unspentKey)
+	unspent, err = unspentRepository.GetUnspentWithKey(ctx, unspentKey)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -231,7 +232,7 @@ func TestConfirmUnspents(t *testing.T) {
 		TxID: "2",
 		VOut: 1,
 	}
-	unspent, err := unspentRepository.GetUnspentForKey(ctx, unspentKey)
+	unspent, err := unspentRepository.GetUnspentWithKey(ctx, unspentKey)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -245,7 +246,7 @@ func TestConfirmUnspents(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	unspent, err = unspentRepository.GetUnspentForKey(ctx, unspentKey)
+	unspent, err = unspentRepository.GetUnspentWithKey(ctx, unspentKey)
 	if err != nil {
 		t.Fatal(err)
 	}
